Revert "Create API for registration (#4)"
This reverts commit 398219c5429a13c2c62918792709d18cf54342ed.
