+++ backend/backend.py
@@ -1,0 +1,13 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+
+@app.route('/')
+def index():
+    return "Hello world"
+
+
+if __name__ == "__main__":
+    app.run(debug=True)
+
