--- requirements.txt
@@ -1,1 +1,0 @@
-Flask==2.2.3
