--- src/config.py
@@ -1,9 +1,0 @@
-from flask import Flask
-
-# Flask app
-app = Flask(__name__)
-
-# Dict with rate limits
-rate_limits = {
-    "default": "6 per minute"
-}
